requirements.txt
Django
elasticsearch
elasticsearch-dsl
peewee
lxml
django-cors-headers
djangorestframework